mybatis config added importplanalarm service tested

+++ src/main/java/com/seohan/mat/Dto/ImportPlanAlarm.java
@@ -8,21 +8,22 @@ import javax.persistence.Id;
 import javax.persistence.IdClass;
 import javax.persistence.Table;
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 public class ImportPlanAlarm   {
     private String      cstcd	;
     private String      itmno	;
     private String      warhs	;
-    private Integer         mqty	;
-    private Integer         bsqty	;
-    private Integer         tsqty	;
-    private Integer         preqty	;
-    private Integer         expqty	;
+    private BigDecimal   mqty	;
+    private BigDecimal         bsqty	;
+    private BigDecimal         tsqty	;
+    private BigDecimal         preqty	;
+    private BigDecimal         expqty	;
     private String      dscrp	;
     private String      cusna	;
 
-    public ImportPlanAlarm(String cstcd, String itmno, String warhs, Integer mqty, Integer bsqty, Integer tsqty, Integer preqty, Integer expqty, String dscrp, String cusna) {
+    public ImportPlanAlarm(String cstcd, String itmno, String warhs, BigDecimal mqty, BigDecimal bsqty, BigDecimal tsqty, BigDecimal preqty, BigDecimal expqty, String dscrp, String cusna) {
         this.cstcd = cstcd;
         this.itmno = itmno;
         this.warhs = warhs;
